extractor-gdelt/requirements.txt
arrow==0.15.4
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
dateparser==0.7.2
google-api-core==1.14.3
google-auth==1.7.1
google-cloud-bigquery==1.22.0
google-cloud-core==1.0.3
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
idna==2.8
moment==0.8.2
numpy==1.17.3
pandas==0.25.3
pika==1.1.0
protobuf==3.11.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
regex==2019.11.1
requests==2.22.0
rsa==4.0
six==1.13.0
times==0.7
tqdm==4.38.0
tzlocal==2.0.0
urllib3==1.25.6
